pyproject.toml
[project]
name = "po-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"  
dependencies = [
    "streamlit>=1.38.0",
    "paddleocr>=3.0",
    "paddlepaddle>=3.1.1",
    "paddleocr[doc-parser]>=3.0", # <-- pulls the PP-StructureV3 deps
    "pymupdf>=1.23.0",
    "pandas>=2.2.0",
    "lxml>=4.9.3",
    "pdfplumber>=0.11.0",
    "ipykernel>=6.30.0",
    "PyMuPDF>=1.26.3",
    "streamlit >=1.50.0",
    "auth >=0.5.3",
    "python-dotenv",
    "plotly",
    "openai >=1.108.1",
    "python-dotenv",
]

